requirements.txt
aiohttp==3.8.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asn1crypto==1.5.1
asttokens==2.0.8
async-timeout==4.0.2
asyncssh==2.13.0
atpublic==3.1.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
configobj==5.0.6
cryptography==36.0.2
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
dpath==2.1.3
dulwich==0.20.50
dvc==2.38.1
dvc-data==0.28.4
dvc-http==2.27.2
dvc-objects==0.14.0
dvc-render==0.0.15
dvc-task==0.1.8
dvclive==1.3.0
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.8.0
flatten-dict==0.4.2
flufl.lock==7.1.1
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
funcy==1.17
future==0.18.2
gitdb==4.0.10
GitPython==3.1.29
grandalf==0.6
greenlet==1.1.3.post0
hydra-core==1.3.1
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
ipykernel==6.19.4
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
iterative-telemetry==0.0.6
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.5
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
nanotime==0.5.2
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
omegaconf==2.3.0
oscrypto==1.3.0
packaging==22.0
pandas==1.5.0
pandasql==0.7.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.15.0
pydot==1.4.2
pygit2==1.11.1
Pygments==2.13.0
pygtrie==2.5.0
PyJWT==2.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.0
scipy==1.9.3
scmrepo==0.1.4
Send2Trash==1.8.0
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
snowflake-connector-python==2.8.0
snowflake-sqlalchemy==1.4.2
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
stack-data==0.5.1
statsmodels==0.13.5
tabulate==0.9.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.12
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
yarl==1.8.2
zc.lockfile==2.0
zipp==3.11.0
